+++ tilesets/tilesets.js
@@ -3,26 +3,34 @@
         {
             folder: "default",
             name: 'Original',
+            color: 'cyan'
         }
         ,{
             folder: "lostgift",
             name: 'Lost Gifts',
+            color: 'orange',
             tile: true,
             object: true
         }
         /*,{
             folder: "new folder",
             name: 'new name',
+            color: 'hex value',
             tile: true
         }*/
         
-        //["FOLDER NAME", "SHOWN NAME", add which sheet the custom tileset uses]
+        //[
+        //  "FOLDER NAME",
+        //  "SHOWN NAME",
+        //  "color in hex or any css valid color" DONT USE RED OR LIME
+        //  add which sheet the custom tileset uses]
 
 
 
         // ,{
         //     folder: 'custom',
         //     name: 'Modded',
+        //     color: '#ffffff or white',
         //     tile: false,
         //     object: true,
         //     volume: true